Add a spec illustrating the bug fixed with v0.3.1
Prior to v0.3.1, the macro that generated the stack-supporting event
handlers would wrongly register an event handler that would run the
entire stack multiple times. In the given example in this spec,
times_called would have a value of four since two event handlers were
created. Now, the instance variable that stored stacks is removed and we
instead use the closure of the event handler to store the registered
stack, so that only one event handler is created per-stack.